Simplified the DTOs
