+++ Dtos/CommandDto.cs
@@ -1,0 +1,40 @@
+using System.ComponentModel.DataAnnotations;
+
+namespace Commander.Dtos
+{
+    public class CommandCreateDto
+    {
+        [Required]
+        [MaxLength(250)]
+        public string HowTo { get; set; }
+        
+        [Required]
+        public string Line { get; set; }
+        
+        [Required]
+        public string Platform { get; set; }
+    }
+
+    public class CommandReadDto
+    {
+        public int Id { get; set; }
+        
+        public string HowTo { get; set; }
+        
+        public string Line { get; set; }
+        
+    }
+
+    public class CommandUpdateDto
+    {
+        [Required]
+        [MaxLength(250)]
+        public string HowTo { get; set; }
+        
+        [Required]
+        public string Line { get; set; }
+        
+        [Required]
+        public string Platform { get; set; }
+    }
+}
