--- Dtos/CommandReadDto.cs
@@ -1,12 +1,0 @@
-namespace Commander.Dtos
-{
-    public class CommandReadDto
-    {
-        public int Id { get; set; }
-        
-        public string HowTo { get; set; }
-        
-        public string Line { get; set; }
-        
-    }
-}
